+++ R/RKinetica-package.R
@@ -1,0 +1,17 @@
+#' RKinetica
+#'
+#' RKinetica is an R package providing access to Kinetica DB through REST interface.
+#' Kinetica DB is a  distributed, in-memory, GPU-accelerated database that utilizes a
+#' powerful combination of CPUs and GPUs to analyze massive, complex datasets with
+#' millisecond response times.
+#'
+#' At the core of Kinetica is a vectorized, columnar, memory-first database designed for
+#' analytical (OLAP) workloads. Kinetica automatically distributes any workload across CPUs
+#' and GPUs for optimal results.
+#'
+#' For more information on Kinetica DB please visit our web site \url{http://www.kinetica.com}
+#' @docType package
+#' @name RKinetica
+#' @aliases RKinetica RKinetica-package
+#' @keywords package
+NULL