+++ R/zzz.R
@@ -1,0 +1,6 @@
+.onLoad <- function(libname, pkgname) {
+
+}
+.onUnload <- function(libpath) {
+  gc()
+}